backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
redis==5.0.1
APScheduler==3.10.4
stripe==7.7.0
PyJWT==2.8.0
google-generativeai==0.3.2
python-dotenv==1.0.0
eventlet==0.33.3
qrcode[pil]==7.4.2
